chore: Rename application
Rename the application to its proper name: Bank API.
