+++ src/main/kotlin/com/ricardocosta/api/bank/configuration/JacksonConfig.kt
@@ -1,4 +1,4 @@
-package com.ricardocosta.template.configuration
+package com.ricardocosta.api.bank.configuration
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.KotlinModule
