+++ src/main/kotlin/com/ricardocosta/api/bank/dto/HelloDTO.kt
@@ -1,4 +1,4 @@
-package com.ricardocosta.template.dto
+package com.ricardocosta.api.bank.dto
 
 data class HelloDTO(
     val name: String
